bugfix: Fix read_line panic on non-empty buffer
The behavior of the canonical read_line is to append to whatever buffer
is passed into it. At the moment it assumes the buffer is empty and
panics otherwise. This commit adds a check for the buffer being empty,
and appends the string otherwise.

This commit also adds a unit test to prevent regressions.

Fixes #124